backend/requirements.txt
fastapi
uvicorn
sqlmodel
google-cloud-vision
openai
python-multipart
python-dotenv
httpx
